Move serialization error definition into parsl.serialize
This is in expectation of at least one more exception being added
related to serialization as part of ongoing serializer plugin
work.
