+++ parsl/tests/test_error_handling/test_serialization_fail.py
@@ -2,7 +2,7 @@ import pytest
 
 from parsl import python_app
 from parsl.tests.configs.htex_local import fresh_config
-from parsl.executors.errors import SerializationError
+from parsl.serialize.errors import SerializationError
 
 
 def local_config():
